requirements.txt
replicate
requests
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib